commit inicial
This commit contains: 1)index.html: the HTML file, structured with tags, headings, paragraphs, and links. It's the backbone of the webpage. 2)styles.css: a CSS file that provides the visual design, including font styling, layout adjustments, and color schemes. 3)/img/ folder: a directory containing all the images used in the project.